[feat] refactor graph runner
