+++ python/sglang/srt/model_executor/npu_graph_runner.py
@@ -1,4 +1,4 @@
-# Copyright 2025 SGLang Team
+# Copyright 2023-2024 SGLang Team
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -11,76 +11,104 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-"""Run the model with npu graph engine and torch.compile."""
+"""Run the model with npu graph and torch.compile."""
 
 from __future__ import annotations
 
-import bisect
 import inspect
+import logging
 import os
+import threading
 import types
 from contextlib import contextmanager
-from typing import (
-    TYPE_CHECKING,
-    Any,
-    Callable,
-    ContextManager,
-    Generator,
-    Optional,
-    Tuple,
-    Union,
-)
+from typing import TYPE_CHECKING, Any, Callable, Generator, Optional, Union
 
 import torch
 import tqdm
-from networkx.utils.backends import backends
-
 from sglang.srt.distributed import get_tensor_model_parallel_rank
 from sglang.srt.layers.logits_processor import LogitsProcessorOutput
-from sglang.srt.managers.schedule_batch import global_server_args_dict
-from sglang.srt.model_executor.cuda_graph_runner import (
-    DeviceRunnerBase,
-    _to_torch,
-    get_batch_sizes_to_capture,
-    patch_model,
-    set_torch_compile_config,
-)
 from sglang.srt.model_executor.forward_batch_info import (
     CaptureHiddenMode,
     ForwardBatch,
-    ForwardMode,
     PPProxyTensors,
-    enable_num_token_non_padded,
 )
+from sglang.srt.model_executor.graph_runner import GraphRunner
 from sglang.srt.utils import (
     get_available_gpu_memory,
     get_compiler_backend,
-    get_device,
-    get_device_memory_capacity,
-    rank0_log,
 )
+
+logger = logging.getLogger(__name__)
 
 if TYPE_CHECKING:
     from sglang.srt.model_executor.model_runner import ModelRunner
 
 
-class NpuGraphRunner(DeviceRunnerBase):
-    """A NpuGraphRunner runs the forward pass of a model with npu graph engine and torch.compile."""
+class NPUGraphRunner(GraphRunner):
+    """A NPUGraphRunner runs the forward pass of a model with npu graph and torch.compile."""
 
     def __init__(self, model_runner: ModelRunner):
         super().__init__(model_runner)
-        self.enable_cache = os.getenv("ENABLE_TORCH_COMPILE_CACHE", "0") == "1"
-        try:
+        if self.enable_torch_compile:
+            self.enable_cache = os.getenv("ENABLE_TORCH_COMPILE_CACHE", "0") == "1"
             self.warm_up()
-        except RuntimeError as e:
-            raise Exception(
-                f"compile npu graph failed: {e}\n"
-                "Possible solutions:\n"
-                "1. set --mem-fraction-static to a smaller value (e.g., 0.8 or 0.7)\n"
-                "2. set --torch-compile-max-bs to a smaller value (e.g., 16)\n"
-                "3. disable torch compile by not using --enable-torch-compile\n"
-                "Open an issue on GitHub https://github.com/sgl-project/sglang/issues/new/choose \n"
+        else:
+            super().warm_up()
+
+    def _create_device_graph(self):
+        return torch.npu.NPUGraph()
+
+    def _capture_graph(self, graph, pool, stream, run_once_fn):
+        with torch.npu.graph(
+            graph,
+            pool=pool,
+            stream=stream,
+            auto_dispatch_capture=True,
+        ):
+            out = run_once_fn()
+        return out
+
+    def _update_inputs(self, seq_lens):
+        self.graphs[self.bs].update(
+            cpu_update_input=[{"actual_seq_lengths_kv": seq_lens}]
+        )
+
+    def _cache_loc_dtype(self):
+        return torch.int32
+
+    def replay(
+        self,
+        forward_batch: ForwardBatch,
+        skip_attn_backend_init: bool = False,
+        pp_proxy_tensors: Optional[PPProxyTensors] = None,
+    ) -> Union[LogitsProcessorOutput, PPProxyTensors]:
+        if not skip_attn_backend_init:
+            self.replay_prepare(forward_batch, pp_proxy_tensors)
+        else:
+            # In speculative decoding, these two fields are still needed.
+            self.input_ids[: self.raw_num_token].copy_(forward_batch.input_ids)
+            self.positions[: self.raw_num_token].copy_(forward_batch.positions)
+
+        # Replay
+        seq_lens = forward_batch.seq_lens.cpu().tolist() + [0] * (self.bs - self.raw_bs)
+        thread = threading.Thread(target=self._update_inputs, args=(seq_lens,))
+        thread.start()
+        self.graphs[self.bs].replay()
+        thread.join()
+
+        output = self.output_buffers[self.bs]
+        if isinstance(output, LogitsProcessorOutput):
+            return LogitsProcessorOutput(
+                next_token_logits=output.next_token_logits[: self.raw_num_token],
+                hidden_states=(
+                    output.hidden_states[: self.raw_num_token]
+                    if output.hidden_states is not None
+                    else None
+                ),
             )
+        else:
+            assert isinstance(output, PPProxyTensors)
+            return PPProxyTensors({k: v[: self.bs] for k, v in output.tensors.items()})
 
     def prepare_forward_batch(self, bs: int, num_tokens: int) -> ForwardBatch:
         # Graph inputs
@@ -168,50 +196,7 @@ class NpuGraphRunner(DeviceRunnerBase):
         )
         return forward_batch
 
-    def mark_static(
-        self, forward_batch: ForwardBatch, pp_proxy_tensors: PPProxyTensors = None
-    ):
-        def mark_tensor_static(model_input, is_cache=False):
-            if model_input is not None:
-                if isinstance(model_input, torch.Tensor):
-                    torch._dynamo.mark_static(model_input)
-                elif is_cache:
-                    for buffer_per_layer in model_input:
-                        torch._dynamo.mark_static(buffer_per_layer)
-                elif isinstance(model_input, PPProxyTensors):
-                    for pp_out in model_input.tensors.items():
-                        torch._dynamo.mark_static(pp_out)
-                elif isinstance(model_input, tuple):
-                    for value in model_input:
-                        torch._dynamo.mark_static(value)
-                else:
-                    raise ValueError(
-                        f"Unsupported type with mark static: {type(model_input)}"
-                    )
-
-        mark_tensor_static(pp_proxy_tensors)
-        mark_tensor_static(forward_batch.input_ids)
-        mark_tensor_static(forward_batch.positions)
-        mark_tensor_static(forward_batch.input_embeds)
-        mark_tensor_static(forward_batch.out_cache_loc)
-        mark_tensor_static(forward_batch.gathered_buffer)
-        mark_tensor_static(forward_batch.attn_backend.forward_metadata.block_kv_indices)
-        mark_tensor_static(forward_batch.attn_backend.forward_metadata.mc2_mask)
-        try:
-            mark_tensor_static(forward_batch.token_to_kv_pool.k_buffer, is_cache=True)
-            mark_tensor_static(forward_batch.token_to_kv_pool.v_buffer, is_cache=True)
-        except AttributeError as e:
-            mark_tensor_static(forward_batch.token_to_kv_pool.kv_buffer, is_cache=True)
-
     def warm_up(self):
-        if not self.enable_torch_compile:
-            rank0_log(
-                "enable_torch_compile is False, model will run eagerly on npu, this may cause performance loss."
-                "please set --enable-torch-compile"
-            )
-            return
-
-        rank0_log("Warming up npu graph")
         backend = get_compiler_backend()
         if not self.enable_cache:
             self.model_runner.model.compile_forward = torch.compile(
@@ -281,8 +266,8 @@ def {method_name}(self, input_ids, positions, forward_batch, **kwargs):
                     self.pp_size > 1
                     and "pp_proxy_tensors"
                     in inspect.signature(
-                        self.model_runner.model.compile_forward
-                    ).parameters
+                    self.model_runner.model.compile_forward
+                ).parameters
                 ):
                     kwargs["pp_proxy_tensors"] = forward_batch.pp_proxy_tensors
                 self.mark_static(forward_batch, kwargs.get("pp_proxy_tensors"))
@@ -307,7 +292,40 @@ def {method_name}(self, input_ids, positions, forward_batch, **kwargs):
                 self.model_runner.tp_group.barrier()
                 run_once()
 
-        return
+    def mark_static(
+        self, forward_batch: ForwardBatch, pp_proxy_tensors: PPProxyTensors = None
+    ):
+        def mark_tensor_static(model_input, is_cache=False):
+            if model_input is not None:
+                if isinstance(model_input, torch.Tensor):
+                    torch._dynamo.mark_static(model_input)
+                elif is_cache:
+                    for buffer_per_layer in model_input:
+                        torch._dynamo.mark_static(buffer_per_layer)
+                elif isinstance(model_input, PPProxyTensors):
+                    for pp_out in model_input.tensors.items():
+                        torch._dynamo.mark_static(pp_out)
+                elif isinstance(model_input, tuple):
+                    for value in model_input:
+                        torch._dynamo.mark_static(value)
+                else:
+                    raise ValueError(
+                        f"Unsupported type with mark static: {type(model_input)}"
+                    )
+
+        mark_tensor_static(pp_proxy_tensors)
+        mark_tensor_static(forward_batch.input_ids)
+        mark_tensor_static(forward_batch.positions)
+        mark_tensor_static(forward_batch.input_embeds)
+        mark_tensor_static(forward_batch.out_cache_loc)
+        mark_tensor_static(forward_batch.gathered_buffer)
+        mark_tensor_static(forward_batch.attn_backend.forward_metadata.block_kv_indices)
+        mark_tensor_static(forward_batch.attn_backend.forward_metadata.mc2_mask)
+        try:
+            mark_tensor_static(forward_batch.token_to_kv_pool.k_buffer, is_cache=True)
+            mark_tensor_static(forward_batch.token_to_kv_pool.v_buffer, is_cache=True)
+        except AttributeError as e:
+            mark_tensor_static(forward_batch.token_to_kv_pool.kv_buffer, is_cache=True)
 
     @contextmanager
     def get_runner_context(
@@ -316,23 +334,26 @@ def {method_name}(self, input_ids, positions, forward_batch, **kwargs):
         skip_attn_backend_init: bool,
     ) -> Generator[Callable[[bool, PPProxyTensors | None], Any], Any, None]:
         def runner_fn(pp_proxy_tensors: Optional[PPProxyTensors]):
-            kwargs = {}
-            if pp_proxy_tensors is not None:
-                kwargs["pp_proxy_tensors"] = pp_proxy_tensors
+            if self.enable_torch_compile:
+                kwargs = {}
+                if pp_proxy_tensors is not None:
+                    kwargs["pp_proxy_tensors"] = pp_proxy_tensors
 
-            compile_method_name = f"compile_forward_{forward_batch.input_ids.size(0)}bs"
-            compile_forward = (
-                getattr(self.model_runner.model, compile_method_name)
-                if self.enable_cache
-                else self.model_runner.model.compile_forward
-            )
-            with torch.no_grad():
-                return compile_forward(
-                    forward_batch.input_ids,
-                    forward_batch.positions,
-                    forward_batch,
-                    **kwargs,
+                compile_method_name = f"compile_forward_{forward_batch.input_ids.size(0)}bs"
+                compile_forward = (
+                    getattr(self.model_runner.model, compile_method_name)
+                    if self.enable_cache
+                    else self.model_runner.model.compile_forward
                 )
+                with torch.no_grad():
+                    return compile_forward(
+                        forward_batch.input_ids,
+                        forward_batch.positions,
+                        forward_batch,
+                        **kwargs,
+                    )
+            else:
+                return self.replay(forward_batch, skip_attn_backend_init, pp_proxy_tensors)
 
         if not skip_attn_backend_init:
             forward_batch.attn_backend.init_forward_metadata(forward_batch)
